refactor(improvements/promotion): merge inline tests into tests_promotion.rs (#1266)
Inline test module in promotion.rs is moved into existing tests_promotion.rs
as a nested mod promotion_inline {} submodule (preemptive scoping for any
future helper collisions).

- src/improvements/promotion.rs: 491 → 362 LOC (under 400-line rule)
- src/improvements/tests_promotion.rs: 81 → 217 LOC

Pure refactor: zero logic changes, all 12 tests pass.

Closes 13/70 of #1266 (this loop iteration), 22 PRs cumulative since #1285.
