+++ src/improvements/promotion.rs
@@ -359,133 +359,3 @@ fn parse_deferral_directive(raw: &str) -> SimardResult<DeferredImprovement> {
 
     Ok(DeferredImprovement { title, rationale })
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn valid_plan_text() -> String {
-        [
-            "review-id: rev-001",
-            "review-target: benchmark-run",
-            "proposal: Fix flaky test | category=testing | rationale=Reduces CI noise | suggested_change=Add retry logic | evidence=ci-log-42",
-            "approve: Fix flaky test | priority=2 | status=active | rationale=High impact fix",
-        ]
-        .join("\n")
-    }
-
-    #[test]
-    fn parse_valid_plan() {
-        let plan = ImprovementPromotionPlan::parse(&valid_plan_text()).unwrap();
-        assert_eq!(plan.review_id, "rev-001");
-        assert_eq!(plan.review_target, "benchmark-run");
-        assert_eq!(plan.proposals.len(), 1);
-        assert_eq!(plan.approvals.len(), 1);
-        assert!(plan.deferrals.is_empty());
-    }
-
-    #[test]
-    fn parse_missing_review_id_errors() {
-        let raw = [
-            "proposal: X | category=c | rationale=r | suggested_change=s | evidence=e",
-            "approve: X",
-        ]
-        .join("\n");
-        assert!(ImprovementPromotionPlan::parse(&raw).is_err());
-    }
-
-    #[test]
-    fn parse_no_proposals_errors() {
-        let raw = "review-id: rev-001\napprove: X";
-        assert!(ImprovementPromotionPlan::parse(raw).is_err());
-    }
-
-    #[test]
-    fn parse_no_decisions_errors() {
-        let raw = [
-            "review-id: rev-001",
-            "proposal: X | category=c | rationale=r | suggested_change=s | evidence=e",
-        ]
-        .join("\n");
-        assert!(ImprovementPromotionPlan::parse(&raw).is_err());
-    }
-
-    #[test]
-    fn parse_unknown_decision_title_errors() {
-        let raw = [
-            "review-id: rev-001",
-            "proposal: Real title | category=c | rationale=r | suggested_change=s | evidence=e",
-            "approve: Wrong title",
-        ]
-        .join("\n");
-        assert!(ImprovementPromotionPlan::parse(&raw).is_err());
-    }
-
-    #[test]
-    fn parse_with_deferral() {
-        let raw = [
-            "review-id: rev-002",
-            "proposal: Later thing | category=perf | rationale=low impact | suggested_change=optimize | evidence=bench-data",
-            "defer: Later thing | rationale=Not a priority now",
-        ]
-        .join("\n");
-        let plan = ImprovementPromotionPlan::parse(&raw).unwrap();
-        assert_eq!(plan.deferrals.len(), 1);
-        assert_eq!(plan.deferrals[0].title, "Later thing");
-    }
-
-    #[test]
-    fn approval_summaries_format() {
-        let plan = ImprovementPromotionPlan::parse(&valid_plan_text()).unwrap();
-        let summaries = plan.approval_summaries();
-        assert_eq!(summaries.len(), 1);
-        assert!(summaries[0].starts_with("p2"));
-        assert!(summaries[0].contains("[active]"));
-    }
-
-    #[test]
-    fn deferral_summaries_format() {
-        let raw = [
-            "review-id: rev-003",
-            "proposal: X | category=c | rationale=r | suggested_change=s | evidence=e",
-            "defer: X | rationale=Not yet",
-        ]
-        .join("\n");
-        let plan = ImprovementPromotionPlan::parse(&raw).unwrap();
-        let summaries = plan.deferral_summaries();
-        assert_eq!(summaries.len(), 1);
-        assert!(summaries[0].contains("Not yet"));
-    }
-
-    #[test]
-    fn render_review_context_directives_includes_review_id() {
-        let review = ReviewArtifact {
-            review_id: "rev-100".into(),
-            reviewed_at_unix_ms: 0,
-            target_kind: crate::review::ReviewTargetKind::Benchmark,
-            target_label: "suite:scenario".into(),
-            identity_name: "id".into(),
-            session_id: "s".into(),
-            selected_base_type: "bt".into(),
-            topology: "single-process".into(),
-            objective_metadata: "meta".into(),
-            execution_summary: "exec".into(),
-            reflection_summary: "refl".into(),
-            summary: "sum".into(),
-            measurement_notes: vec![],
-            evidence_summary: crate::review::ReviewEvidenceSummary {
-                memory_records: 0,
-                evidence_records: 0,
-                decision_records: 0,
-                benchmark_records: 0,
-                exported_state: "stopped".into(),
-                session_phase: None,
-                failed_signals: vec![],
-            },
-            proposals: vec![],
-        };
-        let output = render_review_context_directives(&review);
-        assert!(output.contains("review-id: rev-100"));
-        assert!(output.contains("review-target: suite:scenario"));
-    }
-}
